refactor(pi-agents-tmux): split browser.ts into focused modules
Redo of PR #54 against current main. Same intent, applied on top of newer
Monitor/dashboard work so the structural cleanup composes with it instead
of conflicting.

Extracts /agents popup concerns from extensions/subagent/browser.ts (1983
lines) into per-concern modules under extensions/subagent/browser/, and a
new neutral extensions/subagent/task-records.ts:

- browser/shared.ts \u2014 modal lock, frame/pad/tab strip, layout helpers.
- browser/agents-tab.ts \u2014 agent list + static Inspector.
- browser/monitor-tree.ts \u2014 MonitorSessionGroup/Row types, group builder,
  tree row builder, tree renderer.
- browser/monitor-session-detail.ts \u2014 selected-session Detail pane.
- browser/monitor-task-detail.ts \u2014 selected-task Detail, trace content
  renderer, renderTraceTabBar, traceViewerItems.
- browser/frontmatter-editor.ts \u2014 YAML/TOML parse + upsert, editor flow,
  post-edit confirmation popup.
- browser/dashboard-integration.ts \u2014 activeDashboardItems,
  dashboardDisplayLabels, appendBgChatMessages.
- browser/trace-viewer.ts \u2014 standalone openTraceViewer popup; shares
  renderTraceTabBar with Monitor.
- task-records.ts \u2014 monitorSessionKey, recordMonitorKind, taskNumberById,
  sortedMonitorRecords, usageSum, status helpers, loadTaskRegistrySync.
  Dashboard imports these directly instead of via browser.ts, removing
  the browser.ts \u2194 dashboard.ts cycle.

browser.ts keeps modal lifecycle, input dispatch, top-level tab
composition, and re-exports every symbol that was public pre-split so
tests/index.ts/agents-command.ts continue to import from ./browser.js
without churn.

Closes #55.

Validation:
- tsc --noEmit on extensions/subagent/*.ts extensions/subagent/browser/*.ts \u2014 clean.
- tsc --noEmit --noUnusedLocals --noUnusedParameters on the new files \u2014 clean.
- npm test (bun test, 66 tests) \u2014 all pass.
- npm pack --dry-run \u2014 succeeds.
- madge --circular \u2014 only pre-existing agents.ts/settings.ts/types.ts
  cycles; the browser.ts \u2194 dashboard.ts cycle from main is gone.
- vstack verify -g @vanillagreen/pi-agents-tmux \u2014 OK on current main
  baseline (post-merge vstack refresh -g will pick this up).

Reviewers: reviewer-structure, reviewer-arch, reviewer-doc,
reviewer-test, reviewer-error \u2014 all approve after follow-up that
(a) reunified renderTraceTabBar in monitor-task-detail.ts and dropped
the duplicate from trace-viewer.ts, (b) dropped the duplicate
loadAgentPaneStatuses from browser.ts in favor of the shared.ts copy,
(c) restored renderMonitorSessionDetail's discovery param type to
{ agents: AgentConfig[] } to match pre-refactor signature.
